refactor out preview component
