+++ codepen-clone/.eslintcache
@@ -1,1 +1,1 @@
-[{"/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/index.tsx":"1","/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/App.tsx":"2","/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/components/code-editor.tsx":"3","/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/plugins/unpkg-path-plugin.ts":"4","/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/plugins/fetch-plugin.ts":"5"},{"size":249,"mtime":1623241596645,"results":"6","hashOfConfig":"7"},{"size":2212,"mtime":1626435364356,"results":"8","hashOfConfig":"7"},{"size":1742,"mtime":1623237859200,"results":"9","hashOfConfig":"7"},{"size":826,"mtime":1624468967693,"results":"10","hashOfConfig":"7"},{"size":2099,"mtime":1625667073896,"results":"11","hashOfConfig":"7"},{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"14"},"ssvdrz",{"filePath":"15","messages":"16","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"17","messages":"18","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"14"},{"filePath":"19","messages":"20","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"14"},{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"14"},"/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/index.tsx",[],["23","24"],"/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/App.tsx",[],"/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/components/code-editor.tsx",[],"/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/plugins/unpkg-path-plugin.ts",[],"/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/plugins/fetch-plugin.ts",[],{"ruleId":"25","replacedBy":"26"},{"ruleId":"27","replacedBy":"28"},"no-native-reassign",["29"],"no-negated-in-lhs",["30"],"no-global-assign","no-unsafe-negation"]
+[{"/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/index.tsx":"1","/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/App.tsx":"2","/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/components/code-editor.tsx":"3","/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/plugins/unpkg-path-plugin.ts":"4","/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/plugins/fetch-plugin.ts":"5","/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/components/preview.tsx":"6"},{"size":249,"mtime":1623241596645,"results":"7","hashOfConfig":"8"},{"size":1482,"mtime":1626437301185,"results":"9","hashOfConfig":"8"},{"size":1742,"mtime":1623237859200,"results":"10","hashOfConfig":"8"},{"size":826,"mtime":1624468967693,"results":"11","hashOfConfig":"8"},{"size":2099,"mtime":1625667073896,"results":"12","hashOfConfig":"8"},{"size":998,"mtime":1626437302041,"results":"13","hashOfConfig":"8"},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},"ssvdrz",{"filePath":"17","messages":"18","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"19","messages":"20","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/index.tsx",[],["27","28"],"/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/App.tsx",[],"/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/components/code-editor.tsx",[],"/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/plugins/unpkg-path-plugin.ts",[],"/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/plugins/fetch-plugin.ts",[],"/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/components/preview.tsx",[],{"ruleId":"29","replacedBy":"30"},{"ruleId":"31","replacedBy":"32"},"no-native-reassign",["33"],"no-negated-in-lhs",["34"],"no-global-assign","no-unsafe-negation"]
